회원가입 완료. 남은것 - NAVER, KAKAO API 라이브러리
